+++ src/jtl/Connector/Shopware/Controller/Image.php
@@ -63,6 +63,18 @@ class Image extends DataController
                         case ImageRelationType::TYPE_PRODUCT:
                             $model = Mmc::getModel('Image');
 
+                            $id = ImageModel::generateId(ImageRelationType::TYPE_PRODUCT, (int) $modelSW['id'], (int) $modelSW['media_id']);
+                            $path = $modelSW['path'];
+                            $foreignKey = IdConcatenator::link(array($modelSW['detailId'], $modelSW['articleID']));
+
+                            $model->setId(new Identity($id));
+                            $model->setRelationType($relationType)
+                                ->setForeignKey(new Identity($foreignKey))
+                                ->setFilename(sprintf('%s://%s%s/%s', $proto, Shopware()->Shop()->getHost(), Shopware()->Shop()->getBasePath(), $path))
+                                ->setRemoteUrl(sprintf('%s://%s%s/%s', $proto, Shopware()->Shop()->getHost(), Shopware()->Shop()->getBasePath(), $path))
+                                ->setSort((int) $modelSW['position']);
+
+                            /*
                             // Parent
                             $id = ImageModel::generateId(ImageRelationType::TYPE_PRODUCT, $modelSW['id'], $modelSW['media']['id']);
                             $path = $modelSW['media']['path'];
@@ -82,6 +94,7 @@ class Image extends DataController
                                 ->setFilename(sprintf('%s://%s%s/%s', $proto, Shopware()->Shop()->getHost(), Shopware()->Shop()->getBasePath(), $path))
                                 ->setRemoteUrl(sprintf('%s://%s%s/%s', $proto, Shopware()->Shop()->getHost(), Shopware()->Shop()->getBasePath(), $path))
                                 ->setSort($modelSW['position']);
+                            */
 
                             $result[] = $model;
                             break;